Revise the whole codebase: pin the lint rules, kill dead code, fix drifted numbers
The build output is byte-identical before and after every change here, checked after each
step, so none of this moves a published figure.

A PINNED LINT CONFIG, because there wasn't one

`ruff check scripts/` used ruff's small default set while `--select ALL` reported hundreds
of findings, and neither was "the" answer. The practical harm was worse than the ambiguity:
a `# noqa: BLE001` comment explaining *why* a broad `except Exception` is deliberate looked
like an unused directive, and `--fix` deleted the explanation. ruff.toml now pins the set
(F, E, W, ARG, B, BLE, C4, PIE, RET, SIM, RUF at 100 columns), the suppressions mean
something again, and check.yml enforces it so this cannot rot.

The wider rule set immediately earned itself:

  B007  a fixture tuple carried 13 values into 12 columns, so one value was silently
        discarded and `releases` took the wrong field. A data bug, invisible to the eye.
  ARG   two dead function parameters — `models` in _cohorts_by_status, `status` in
        _duration — removed along with their call-site arguments.

DEAD CODE

`load.superseded_snapshots()` was defined, referenced nowhere, and duplicated what
`collect_common.prune_superseded` already does. Found by walking the AST for module-level
names whose only occurrence is their own definition, which is a check lint does not do.

`attach_github_counts` ran TWICE per build — once in build_all for the public/private
split, once inside repositories.build — two passes over the collected frame for identical
output. Attached once and shared, and the per-row name lookup hoisted out of the per-column
loop, which was resolving the same six indexes over and over.

ONE MORE INSTANCE OF THE BUG CI CAUGHT

The IndexError from `col()` returning an empty Series was not confined to kpis: the
publications affiliation mask had the same shape, and would have raised on a table with no
`ersilia_affiliation` column. Guarded — and then generalised into a test: dropping each of
45 columns from publications, repositories and models one at a time, and running every
section builder, now raises nothing. Nine degradation paths verified too, including a fresh
clone with no collected data at all.

NUMBERS IN DOCSTRINGS HAD DRIFTED

The collector picked up new repositories, so several quoted figures were stale — 384 public
repositories is now 386, 241 model repositories is 243. And one had been invalidated by my
own change: collaboration breadth said "29 of 42 papers" when the metric now covers
affiliated papers only, where it is 20 of 25. All recomputed and corrected, and the one
claim most likely to drift again — the external-PR counts — replaced with a pointer to the
build-time caption rather than a number.

Everything asserted in a docstring was re-verified against the data: 237 of 243 models
resolve in all three sources, all 243 identifiers match MODEL_RE, exactly 10 repositories
begin with "eos" and are not models, and eos6ru5 and eos6wdw are still the two absent from
public GitHub.

Also: 17 over-length lines reflowed (all pre-existing), redundant int() casts and list
concatenations simplified, and two blind excepts that had no explanation given one.

+++ scripts/site_data/insights.py
@@ -14,7 +14,7 @@ prose, so the caller owns the wording.
 def num(value):
     """Thousands-separated integer."""
     try:
-        return "{:,}".format(int(round(float(value))))
+        return "{:,}".format(round(float(value)))
     except (TypeError, ValueError):
         return str(value)
